requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
paramiko==3.3.1
requests==2.31.0
geoip2==4.7.0
psutil==5.9.5
cryptography==41.0.4
pyyaml==6.0.1
sqlite3
asyncio
threading
socket
subprocess
json
datetime
logging
ipaddress
